-Alteração para de-serialize dos testes de POST (utilizando verbatim)
-Adicionado estrutura para input data via json file (Utils.RetornaStringJson(jsonFilePath);) em Utils

+++ EcommerceBackend/Models/Order/ModelFee.cs
@@ -8,7 +8,7 @@ namespace DemoRestSharp.models.Order
 {
     public class ModelFee
     {
-        public int price { get; set; }
+        public double price { get; set; }
  
     }
 }
